+++ instafood_frontend/src/authentication/forgotPassword.js
@@ -1,38 +1,13 @@
-import {auth} from "../firebaseConf";
-import {sendPasswordResetEmail} from "firebase/auth";
-import {useState} from "react";
-import {useNavigate, Link} from "react-router-dom";
+import {Link} from "react-router-dom";
+import useAuth from "./authLogic";
 
 export default function ForgotPassword () {
-    const [email, setEmail] = useState("");
 
-    const navigate = useNavigate();
-
-    const handleSendEmail = (e) => {
-        e.preventDefault();
-
-        const continueUrl = "http://localhost:3000/";
-
-        const actionCodeSettings = {
-            url: continueUrl,
-            handleCodeInApp: true,
-        };
-
-        sendPasswordResetEmail(auth, email, actionCodeSettings)
-            .then(() => {
-                alert("Password reset email sent.");
-                navigate("/"); 
-            })
-            .catch((error) => {
-                if (error.code === "auth/invalid-email") {
-                    alert("Invalid email address.");
-                }
-
-                if (error.code === "auth/user-not-found") {
-                    alert("You do not have an account. Please create an account.");
-                }
-            });
-    }
+    const {
+        email,
+        setEmail,
+        handleSendPasswordResetEmail
+    } = useAuth();
 
     return (
         <div>
@@ -42,7 +17,7 @@ export default function ForgotPassword () {
                 value={email}
                 onChange={(e) => setEmail(e.target.value)}
             />
-            <button onClick={handleSendEmail}>Send Password Reset Email</button>
+            <button onClick={handleSendPasswordResetEmail}>Send Password Reset Email</button>
             <br />
             <Link to="/">Back to Login</Link>
         </div>
